First Commit
OpenOffice is a free, open-source, office suite that includes a word processing component, a multimedia powerpoint presentation component, a database component akin to Microsoft Excel, and more. While a powerful alternative to paid office suites like Microsoft Office, OpenOffice is outdated, requiring substantial UI overhauls and feature updates for a more streamlined, usable, and a more desirable product.
One of the software programs included in the suite is OpenOffice Write, a word processing software that provides a large number of formatting and document editing capabilities at no cost. It is often criticized as being cluttered and difficult to navigate, thus not meeting its user base’s expectations for a modern, streamlined layout.